Use basic AJAX instead of collections or models
**it's a massive commit**

*Background:*
REST is not in our case.
The server side has some algorithms on the models
(e.g. when deleting a directory, it deletes recursively).

*Solution:*
We apply AJAX and do non RESTfully.

That implements easily the following features:

* create root directory
* create a directory in a directory
* update a directory's name
* delete a directory (recursively) -> attention: not deleting dir-dir relation, which is yet to be done.

What's more, since it's not a REST way, no collection or model is needed.
We used it only when we for context of templates. That would avoid Backbone's black magics.
Well, more documents need to be read.